Create User API endpoints with validation
- Installed zod for form and request data validation
- Created zod validation schemas for User fields except for timestamps and email verification fields
- Create validation functions for APIs that return NextResponses on error
- Created CRUD API endpoints for User resources, bulk and single actions.
- Created soft-delete and restore endpoints for User

+++ app/api/user/route.ts
@@ -1,5 +1,7 @@
 import { Prisma } from "@/generated/prisma/client";
 import errorHandler from "@/libs/api/errorHandler";
+import { validateIdList } from "@/libs/api/user/validateIdList";
+import validateNewUser from "@/libs/api/user/validateNewUser";
 import { prisma } from "@/libs/db/prisma";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -8,16 +10,38 @@ export async function GET(
 ) {
     const searchParams = request.nextUrl.searchParams;
     
-    const keyword = searchParams.get('keyword') || null;
+    const keyword = searchParams.get('keyword');
     const deleted = searchParams.get('deleted') === 'true' ? true : false;
 
-    const pageParam = searchParams.get('page') || "";
-    let page = Number.isInteger(+pageParam) ? Number(pageParam) : 1;
-    page = (page < 0) ? 1 : page;
+    const intOrDefault = (param: string | null, defaultValue: number): number => {
+        if (param === null) return defaultValue;
+        const num = Number(param);
+        return Number.isInteger(num) && num > 0 ? num : defaultValue;
+    }
 
-    const pageSizeParam = searchParams.get('pageSize') || "";
-    let pageSize = Number.isInteger(+pageSizeParam) ? Number(pageSizeParam) : 10;
-    pageSize = (pageSize < 1) ? 10 : pageSize;
+    const page = intOrDefault(searchParams.get('page'), 1);
+    const pageSize = intOrDefault(searchParams.get('pageSize'), 10);
+
+    const generateGetFilters = (
+        keyword: string | null, 
+        deleted: boolean
+    ): Prisma.UserWhereInput => {
+        const filters: Prisma.UserWhereInput = {}
+
+        if(keyword) {
+            filters.OR = [
+                { username: { contains: keyword }},
+                { email: { contains: keyword }}
+            ]
+        }
+        
+        filters.deletedAt = null
+        if(deleted) {
+            filters.deletedAt = { not: null }
+        }
+
+        return filters;
+    }
 
     try {
         const filters = generateGetFilters(keyword, deleted);
@@ -38,39 +62,20 @@ export async function GET(
     }
 }
 
-function generateGetFilters(
-    keyword: string | null, 
-    deleted: boolean
-): Prisma.UserWhereInput {
-    const filters: Prisma.UserWhereInput = {}
-
-    if(keyword) {
-        filters.OR = [
-            { username: { contains: keyword }},
-            { email: { contains: keyword }}
-        ]
-    }
-    
-    filters.deletedAt = null
-    if(deleted) {
-        filters.deletedAt = { not: null }
-    }
-
-    return filters;
-}
-
 export async function POST(
     request: NextRequest
 ): Promise<NextResponse> {
     // NOTE: Handle email verification in the future
     
     const body = await request.json();
-    const { id, createdAt, updatedAt, deletedAt, emailVerified, verificationToken, tokenExpiresAt, ...user } = body;
+    const validatedUserData = validateNewUser(body);
+
+    if(!validatedUserData.valid) { return validatedUserData.response }
 
     try {
         const newUser = await prisma.user.create({
             data: { 
-                ...user,
+                ...validatedUserData.data,
                 id: 0,
                 emailVerified: false,
                 verificationToken: null,
@@ -95,12 +100,14 @@ export async function PATCH(
     request: NextRequest
 ): Promise<NextResponse> {
     const body = await request.json();
-    const ids: number[] = body.ids || [];
+    const validatedIds = validateIdList(body.ids);
+
+    if(!validatedIds.valid) { return validatedIds.response }
 
     try {
         const deleted = await prisma.user.updateMany({
             where: {
-                id: { in: ids },
+                id: { in: validatedIds.data },
                 deletedAt: null
             },
             data: {
@@ -122,12 +129,14 @@ export async function DELETE(
     request: NextRequest
 ): Promise<NextResponse> {
     const body = await request.json()
-    const ids: number[] = body.ids || [];
+    const validatedIds = validateIdList(body.ids);
 
+    if(!validatedIds.valid) { return validatedIds.response }
+    
     try {
         const deleted = await prisma.user.deleteMany({
             where: {
-                id: { in: ids }
+                id: { in: validatedIds.data }
             }
         })
